Todophase02-backend/requirements.txt
fastapi
sqlmodel
uvicorn[standard]
pydantic-settings
python-jose[cryptography]
python-multipart
psycopg2-binary
passlib==1.7.4
bcrypt==3.2.0
python-dotenv
alembic